Support Cloudflare and Google resolvers
Enhance parsing algorithm by always parsing the query packets from the front so we can properly ignore additional fields we don't care about

Correct response logic so that the response includes an accurate copy of the query the client sent. Previously it would always say the client queried for an A record even if it knew it was responding for an AAAA request.

Move logging toggle logic into log function and cleanup the scattered filters. Log to /var/log/dns for now.

Correct some comments